Suraksha_Electron_App/PolicyVault-Nexus-Real-Time-RBI-Compliant-Data-Gateway/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
cffi==1.17.1
click==8.2.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==45.0.5
Cython==3.1.2
exceptiongroup==1.3.0
fastapi==0.116.1
ffx==2.1.0
FPE==0.1.2
h11==0.16.0
idna==3.10
joblib==1.5.1
numpy==2.2.6
pandas==2.3.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyffx==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
supabase==2.3.4
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0